pyproject.toml
[project]
name = "pdf-ocr"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "camelot-py>=1.0.0",
    "pdfplumber>=0.11.0",
    "docling>=2.82.0",
    "marker-pdf[full]>=1.10.2",
    "odfpy>=1.4.1",
    "opencv-python-headless>=4.10.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pdf2image>=1.17.0",
    "tabulate>=0.10.0",
    "tqdm>=4.67.3",
]
